+++ src/Settings.ts
@@ -14,7 +14,8 @@ export enum Setting {
     ConfigPowerShellWithDocker,
     UseExternalTranslationTool,
     ReplaceSelfClosingXlfTags,
-    SearchOnlyXlfFiles    
+    SearchOnlyXlfFiles,
+    MatchTranslation
 }
 
 export class Settings {
@@ -47,6 +48,7 @@ export class Settings {
         this.SettingCollection[Setting.UseExternalTranslationTool] = this.config.get('UseExternalTranslationTool') ? this.config.get('UseExternalTranslationTool') : false;
         this.SettingCollection[Setting.ReplaceSelfClosingXlfTags] = this.config.get('ReplaceSelfClosingXlfTags');
         this.SettingCollection[Setting.SearchOnlyXlfFiles] = this.config.get('SearchOnlyXlfFiles');
+        this.SettingCollection[Setting.MatchTranslation] = this.config.get('MatchTranslation');
     }
 
     private static getAppSettings(ResourceUri?: vscode.Uri) {